fix: closest_point() now finds point outside polygon
The line from point inside poly to the closest point found on a edge
of the polygon is extended by a given length. Also handles fringe
cases where the point on the edge _is_ a point of the edge. You
then can end up extending the line into the polygon and not outside.
Using the bisector vector of the edges adjacent on the edge point
solves this.

Resolves: #15