Fix: ccw() floating point representation error.
Due to the way floating point numbers are represented, there is only
15-17 significant decimal digits of precision, so in some cases ccw
would evaluate 3 points as not collinear, when they should be.

This is fixed by rounding the area calculation to a fixed tolerance.
I have set this to 13. In the future may allow users to set this
through the API.

Resolves: #19
See also: #20