Fix: shortest_path problem when two equal solutions.
When heappop encounters two items with same key, it moves to evaluate
next item, which in this case is the two Points (i.e. two points have
equal distance). Points are not comparable without a __lt__ function,
but for some reason it works with Python 2.

Add __lt__ comparing the hashes of the Points, which is a simple
way to randomise which Point/solution is chosen.

Resolves: #22