Refactor open_edges into its own class, OpenEdges
Instead of having EdgeKey objects in a list with separate insort
and bisect functions, create a OpenEdges class with insert and
delete methods. There is no need to store p1 and p2 with each
open_edges edge, that was a relic from when open_edges was stored
as a AVL Tree.

This improves code readability and reduces the memory size of
open_edges.